Add wglob-cli to examples
Need to exercise the wchar_t flavor of the library with wglob in testing.
This change adds wglob-cli.cc to examples which converts utf-8 input arguments to wchar_t and invokes glob::wglob API. The tool built from this example code in Xcode is wglob and can be used to test the wchar_t code path like this:
```
tests/glob-test.sh path/to/wglob
```

Renamed `string-match-cli.cc` to `glob-cli.cc` for simplicity and consistency.
cmake builds glob-cli and wglob-cli tools respectively.
