+++ examples/glob.xcodeproj/xcshareddata/xcschemes/glob.xcscheme
@@ -58,7 +58,7 @@
          </CommandLineArgument>
          <CommandLineArgument
             argument = "**/*.cpp"
-            isEnabled = "YES">
+            isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
             argument = "&apos;&apos;"
@@ -66,7 +66,7 @@
          </CommandLineArgument>
          <CommandLineArgument
             argument = "bar/baz/foo.cpp"
-            isEnabled = "YES">
+            isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
             argument = "*.{h,hp,hpp,hxx,c,cc,cp,cpp,cxx}"
@@ -157,12 +157,36 @@
             isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
-            argument = "https://**.google.com"
+            argument = "*.google.com"
             isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
-            argument = "https://foo.bar.google.com"
+            argument = "foo.bar.google.com"
             isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "test*"
+            isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "testa"
+            isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "*.pdf"
+            isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "file.txt"
+            isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "!(foo|bar).txt"
+            isEnabled = "YES">
+         </CommandLineArgument>
+         <CommandLineArgument
+            argument = "baz.txt"
+            isEnabled = "YES">
          </CommandLineArgument>
       </CommandLineArguments>
    </LaunchAction>
